+++ src/cli/index.tsx
@@ -1,13 +1,17 @@
 #!/usr/bin/env node
 
-import { render } from "ink";
-import React from "react";
 import yargs from "yargs";
+import { JSDOM } from "jsdom";
+
+import { dump } from "../tools/dump/dump";
 import { runForwardingServer } from "../tools/proxy/runForwardingServer";
-import { App } from "./App";
 
 // eslint-disable-next-line @typescript-eslint/no-var-requires
 const { hideBin } = require("yargs/helpers");
+
+const jsdom = new JSDOM();
+
+global.DOMParser = jsdom.window.DOMParser;
 
 yargs(hideBin(process.argv))
   .command(
@@ -23,26 +27,16 @@ yargs(hideBin(process.argv))
           type: "string",
           describe: "Path to store the extracted resources",
         })
-        .demandOption(["url"], "Provide a url to the external variables")
         .demandOption(
           ["location"],
           "Provide a location to store the extracted resources"
         );
     },
-    (options: { url: string; location: string }) => {
-      render(
-        <App
-          externalVariablesUrl={options.url}
-          steps={{
-            figureData: true,
-            figureMap: true,
-            furniData: true,
-            figureAssets: true,
-            furniAssets: true,
-          }}
-          location={options.location}
-        />
-      );
+    (options: { url?: string; location: string }) => {
+      dump({
+        externalVariables: options.url,
+        downloadPath: options.location,
+      }).catch(console.error);
     }
   )
   .command(
@@ -67,6 +61,14 @@ yargs(hideBin(process.argv))
           default: false,
           describe: "Sends the length integer as a prefix to the message",
         })
+        .option("key", {
+          type: "string",
+          describe: "Path to key for secure websocket"
+        })
+        .option("cert", {
+          type: "string",
+          describe: "Certificate for secure websocket"
+        })
         .option("debug", {
           type: "boolean",
           describe: "Run in debug mode",
@@ -84,6 +86,8 @@ yargs(hideBin(process.argv))
       port: number;
       prependLengthPrefix: boolean;
       debug: boolean;
+      cert?: string;
+      key?: string;
       targetHost?: string;
     }) => {
       runForwardingServer({
@@ -92,6 +96,8 @@ yargs(hideBin(process.argv))
         debug: options.debug,
         prependLengthPrefix: options.prependLengthPrefix,
         targetHost: options.targetHost,
+        keyPath: options.key,
+        certPath: options.cert,
       });
     }
   )