+++ src/objects/hitdetection/HitTexture.ts
@@ -1,21 +1,30 @@
 import * as PIXI from "pixi.js";
 import { applyTextureProperties } from "../../util/applyTextureProperties";
+import { loadImageFromBlob } from "../../util/loadImageFromBlob";
+import { HitSprite } from "./HitSprite";
 
 export class HitTexture {
   private _texture: PIXI.Texture;
-  private _image: HTMLImageElement;
   private _cachedHitmap: Uint32Array | undefined;
 
   public get texture() {
     return this._texture;
   }
 
-  constructor(image: HTMLImageElement) {
-    if (!image.complete) throw new Error("Image not loaded.");
-
-    this._image = image;
-    this._texture = PIXI.Texture.from(image);
+  constructor(texture: PIXI.Texture) {
+    this._texture = texture;
     applyTextureProperties(this._texture);
+  }
+
+  static async fromSpriteSheet(spritesheet: PIXI.Spritesheet, name: string) {
+    const texture = spritesheet.textures[name];
+    return new HitTexture(texture);
+  }
+
+  static async fromBlob(blob: Blob) {
+    const url = await loadImageFromBlob(blob);
+
+    return HitTexture.fromUrl(url);
   }
 
   static async fromUrl(imageUrl: string) {
@@ -41,7 +50,9 @@ export class HitTexture {
       image.onerror = (value) => reject(value);
     });
 
-    return new HitTexture(image);
+    const texture = PIXI.Texture.from(image);
+
+    return new HitTexture(texture);
   }
 
   hits(
@@ -60,8 +71,8 @@ export class HitTexture {
     const baseTexture = this._texture.baseTexture;
     const hitmap = this._getHitMap();
 
-    const dx = Math.round(x * baseTexture.resolution);
-    const dy = Math.round(y * baseTexture.resolution);
+    const dx = Math.round(this._texture.orig.x + x * baseTexture.resolution);
+    const dy = Math.round(this._texture.orig.y + y * baseTexture.resolution);
     const ind = dx + dy * baseTexture.realWidth;
     const ind1 = ind % 32;
     const ind2 = (ind / 32) | 0;
@@ -70,10 +81,12 @@ export class HitTexture {
 
   private _getHitMap() {
     if (this._cachedHitmap == null) {
-      this._cachedHitmap = generateHitMap(this._image);
+      this._cachedHitmap = generateHitMap(
+        (this._texture.baseTexture.resource as any).source
+      );
     }
 
-    return this._cachedHitmap;
+    return this._cachedHitmap ?? new Uint8ClampedArray();
   }
 }