+++ src/util/parseTileMap.ts
@@ -2,14 +2,17 @@ import { TileType } from "../types/TileType";
 import { getTileInfo } from "./getTileInfo";
 import { ColumnWall, getColumnWalls } from "./tilemap/getColumnWalls";
 import { getRowWalls, RowWall } from "./tilemap/getRowWalls";
+import { padTileMap } from "./tilemap/padTileMap";
+
+export type ParsedTileWall = {
+  type: "wall";
+  kind: "colWall" | "rowWall" | "innerCorner" | "outerCorner";
+  height: number;
+  hideBorder?: boolean;
+};
 
 export type ParsedTileType =
-  | {
-      type: "wall";
-      kind: "colWall" | "rowWall" | "innerCorner" | "outerCorner";
-      height: number;
-      hideBorder?: boolean;
-    }
+  | ParsedTileWall
   | { type: "tile"; z: number }
   | { type: "hidden" }
   | { type: "stairs"; kind: 0 | 2; z: number }
@@ -34,7 +37,7 @@ export function parseTileMap(
 } {
   const wallInfo = new Walls(getRowWalls(tilemap), getColumnWalls(tilemap));
 
-  assertTileMapHasPadding(tilemap);
+  padTileMap(tilemap);
 
   const result: ParsedTileType[][] = tilemap.map((row) =>
     row.map(() => ({ type: "hidden" as const }))
@@ -227,36 +230,3 @@ class Walls {
     return this._colWalls.get(`${x}_${y}`);
   }
 }
-
-function assertTileMapHasPadding(tilemap: TileType[][]) {
-  if (tilemap.length < 1) throw new Error("Tilemap has no rows");
-
-  let doorCount = 0;
-
-  for (let y = 0; y < tilemap.length; y++) {
-    const row = tilemap[y];
-
-    if (row.length < 1) throw new Error("Tilemap row was empty.");
-
-    if (row[0] !== "x") {
-      if (doorCount > 0) throw new Error(paddingErrorMessage(`row ${y}`));
-      doorCount++;
-    }
-  }
-
-  for (let x = 0; x < tilemap[0].length; x++) {
-    const cell = tilemap[0][x];
-
-    if (cell !== "x") throw new Error(paddingErrorMessage(`column ${x}`));
-  }
-}
-
-const paddingErrorMessage = (text: string) => `No padding for ${text}
-- There should be one 'x' at the beginning of each row
-- There should be a full row of 'x' as the first row
-
-Please ensure that the tilemap is padded like the following example:
-xxx
-xoo
-xoo
-`;